Allows to specify graph properties by hostgroup and hostname
Sometimes we have same servicename and check_command for two different
checks (different perfdata, eg. iops for windows and linux hosts using
nrpe). This change adds the possibility to separate graphs my hostgroup,
and hostname. The fields dashboard and panelId are not required anymore,
and the module wont display anything if an override or the variables
don't have values.